chore: merge/move related tests
